Move downloader util to new network mod in burn-common
