+++ burn-core/Cargo.toml
@@ -57,9 +57,10 @@ doc = [
     "burn-wgpu/doc",
 ]
 dataset = ["burn-dataset"]
+network = ["burn-common/network"]
 sqlite = ["burn-dataset?/sqlite"]
 sqlite-bundled = ["burn-dataset?/sqlite-bundled"]
-vision = ["burn-dataset?/vision"]
+vision = ["burn-dataset?/vision", "burn-common/network"]
 
 wasm-sync = ["burn-tensor/wasm-sync", "burn-common/wasm-sync"]
