+++ burn-core/src/data/mod.rs
@@ -7,3 +7,9 @@ pub mod dataloader;
 pub mod dataset {
     pub use burn_dataset::*;
 }
+
+/// Network module.
+#[cfg(feature = "network")]
+pub mod network {
+    pub use burn_common::network::*;
+}
