+++ burn-dataset/Cargo.toml
@@ -21,7 +21,7 @@ fake = ["dep:fake"]
 sqlite = ["__sqlite-shared", "dep:rusqlite"]
 sqlite-bundled = ["__sqlite-shared", "rusqlite/bundled"]
 
-vision = ["dep:flate2", "dep:globwalk", "dep:image", "dep:indicatif", "dep:reqwest", "dep:tokio"]
+vision = ["dep:flate2", "dep:globwalk", "dep:burn-common"]
 
 # internal
 __sqlite-shared = [
@@ -33,20 +33,21 @@ __sqlite-shared = [
 ]
 
 [dependencies]
+burn-common = { path = "../burn-common", version = "0.13.0", optional = true, default-features = false, features = [
+  "network",
+] }
 csv = { workspace = true }
 derive-new = { workspace = true }
 dirs = { workspace = true }
 fake = { workspace = true, optional = true }
 flate2 = { workspace = true, optional = true }
 gix-tempfile = { workspace = true, optional = true }
-globwalk = { workspace = true, optional = true}
+globwalk = { workspace = true, optional = true }
 hound = { workspace = true, optional = true }
 image = { workspace = true, optional = true }
-indicatif = { workspace = true, optional = true }
 r2d2 = { workspace = true, optional = true }
 r2d2_sqlite = { workspace = true, optional = true }
 rand = { workspace = true, features = ["std"] }
-reqwest = { workspace = true, optional = true }
 rmp-serde = { workspace = true }
 rusqlite = { workspace = true, optional = true }
 sanitize-filename = { workspace = true }
@@ -57,7 +58,6 @@ strum = { workspace = true }
 strum_macros = { workspace = true }
 tempfile = { workspace = true }
 thiserror = { workspace = true }
-tokio = { workspace = true, optional = true }
 
 [dev-dependencies]
 rayon = { workspace = true }
