+++ burn-dataset/src/vision/mod.rs
@@ -1,4 +1,3 @@
-mod downloader;
 mod image_folder;
 mod mnist;
