+++ burn/Cargo.toml
@@ -58,6 +58,9 @@ wgpu = ["burn-core/wgpu"]
 tch = ["burn-core/tch"]
 candle = ["burn-core/candle"]
 
+# Network utils
+network = ["burn-core/network"]
+
 # Experimental
 experimental-named-tensor = ["burn-core/experimental-named-tensor"]
